move TrimData to trim_application package
