+++ src/main/java/net/majo24/naturally_trimmed/config/Config.java
@@ -4,7 +4,7 @@ import net.majo24.naturally_trimmed.NaturallyTrimmed;
 import net.majo24.naturally_trimmed.config.backend.ConfigManager;
 import net.majo24.naturally_trimmed.config.backend.annotations.Entry;
 import net.majo24.naturally_trimmed.config.backend.annotations.SubConfig;
-import net.majo24.naturally_trimmed.TrimData;
+import net.majo24.naturally_trimmed.trim_application.TrimData;
 
 import java.util.ArrayList;
 import java.util.List;
